+++ example/lib/main.dart
@@ -8,7 +8,7 @@ void main() {
 }
 
 class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   State<MyApp> createState() => _MyAppState();
@@ -76,8 +76,15 @@ class _MyAppState extends State<MyApp> {
         floatingActionButton: FloatingActionButton(
           child: const Icon(Icons.file_download),
           onPressed: () async {
-            await FlDownloader.download(urlController.text,
-                fileName: 'teste.pdf');
+            final permission = await FlDownloader.requestPermission();
+            if (permission == StoragePermissionStatus.granted) {
+              await FlDownloader.download(
+                urlController.text,
+                fileName: 'teste.pdf',
+              );
+            } else {
+              debugPrint('Permission denied =(');
+            }
           },
         ),
       ),