+++ lib/src/fl_downloader_src.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:typed_data';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+
+import 'storage_permission_status.dart';
 
 part 'progress_entity_src.dart';
 
@@ -10,10 +12,12 @@ class FlDownloader {
   static const MethodChannel _channel = MethodChannel(
     'dev.inceptusp.fl_downloader',
   );
-  static final StreamController<Progress> _progressStream =
+
+  static StreamController<bool> _permissionStatusStream = StreamController();
+  static final StreamController<DownloadProgress> _progressStream =
       StreamController.broadcast();
 
-  static Stream<Progress> get progressStream => _progressStream.stream;
+  static Stream<DownloadProgress> get progressStream => _progressStream.stream;
 
   /// Initializes the plugin and open the stream to listen to download progress
   static Future initialize() async {
@@ -21,15 +25,65 @@ class FlDownloader {
       if (call.method == 'notifyProgress') {
         final map = call.arguments as Map;
         _progressStream.add(
-          Progress._fromMap(<String, dynamic>{...map}),
+          DownloadProgress._fromMap(<String, dynamic>{...map}),
         );
+      }
+      if (call.method == 'onRequestPermissionResult') {
+        final result = call.arguments as bool;
+        _permissionStatusStream.add(result);
+        _permissionStatusStream.close();
+        _permissionStatusStream = StreamController();
       }
       return Future.value(null);
     });
   }
 
+  /// Requests storage permission on Android
+  ///
+  /// If you app supports Android 9 or lower, the call to request storage permission
+  /// is mandatory. Aways returns [StoragePermissionStatus.granted] on Android 10 or higher and on iOS.
+  static Future<StoragePermissionStatus> requestPermission() async {
+    if (Platform.isIOS) return StoragePermissionStatus.granted;
+    try {
+      final status = await _channel.invokeMethod<bool>(
+        'checkStoragePermission',
+      );
+      if (status == true) return StoragePermissionStatus.granted;
+    } catch (e) {
+      debugPrint(e.toString());
+      return StoragePermissionStatus.unknown;
+    }
+
+    try {
+      bool? permissionStatus;
+      await _channel.invokeMethod<bool>('requestStoragePermission');
+      await for (bool event in _permissionStatusStream.stream) {
+        permissionStatus = event;
+      }
+      if (permissionStatus == null) return StoragePermissionStatus.unknown;
+      if (permissionStatus) {
+        return StoragePermissionStatus.granted;
+      } else {
+        return StoragePermissionStatus.denied;
+      }
+    } catch (e) {
+      debugPrint(e.toString());
+      return StoragePermissionStatus.unknown;
+    }
+  }
+
   /// Create and starts a downlaod task on a local URLSession on iOS or
   /// on the system download manager on Android
+  ///
+  /// Returns the id of the download task
+  ///
+  /// If a fileName is not provided, the file name will be extracted from the url and
+  /// if the name extracted from the url contains forbidden characters, this characters
+  /// will be replaced by a dash (-). The list of forbidden characters are:
+  /// ```bash
+  /// # % & { } \ < > * ? / $ ! ' " : @ + ` | =
+  /// ```
+  /// (which covers all characters that are not allowed in most file systems)
   static Future<int> download(
     String url, {
     Map<String, String>? headers,