+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* Updated Dart to 2.18 and Flutter 3.3
+* Renamed `Progress` class to `DownloadProgress` to avoid collision with other classes or plugins
+* Added a explicit documentation related to return of download()
+* Added a method to request Storage Access permition on Android 9 or bellow
+* Fixed download on urls that have invalid filename chars
+
 ## 1.0.3
 
 * Fixed Android `divideByZero` exception